sf_mobile_base: refactor vuex initialization
a
