+++ shopfloor_mobile_base/static/wms/src/main.js
@@ -16,6 +16,7 @@ import {color_registry} from "./services/color_registry.js";
 import {auth_handler_registry} from "./services/auth_handler_registry.js";
 import {Odoo, OdooMocked} from "./services/odoo.js";
 import VueSuperMethod from "./lib/vue-super-call.js";
+import store from "./store/index.js";
 
 Vue.prototype.$super = VueSuperMethod;
 
@@ -53,6 +54,7 @@ config_registry.add("authenticated", {default: false, reset_on_clear: true});
 new Vue({
     i18n,
     router: router,
+    store,
     vuetify: new Vuetify({
         theme: {
             themes: color_registry.get_themes(),
