+++ shopfloor_mobile_base/static/wms/src/store/index.js
