+++ shopfloor_mobile_base/templates/assets.xml
@@ -47,6 +47,16 @@
             t-attf-src="/shopfloor_mobile_base/static/wms/src/lib/vue.js?v=#{app_version}"
         />
         <script
+            t-if="not debug"
+            id="script_vuex_min"
+            t-attf-src="/shopfloor_mobile_base/static/wms/src/lib/vuex.min.js?v=#{app_version}"
+        />
+        <script
+            t-if="debug"
+            id="script_vuex"
+            t-attf-src="/shopfloor_mobile_base/static/wms/src/lib/vuex.js?v=#{app_version}"
+        />
+        <script
             id="script_vue_router"
             t-attf-src="/shopfloor_mobile_base/static/wms/src/lib/vue-router.js?v=#{app_version}"
         />
