+++ shopfloor_shopinvader_mobile_base/static/wms/src/products.js
@@ -7,7 +7,7 @@
 import {process_registry} from "/shopfloor_mobile_base/static/wms/src/services/process_registry.js";
 import {translation_registry} from "/shopfloor_mobile_base/static/wms/src/services/translation_registry.js";
 import ProductModule from "./store/modules/product_store.js";
-import store from "./store/index.js";
+import store from "/shopfloor_mobile_base/static/wms/src/store/index.js";
 
 const products_module = new ProductModule();
 store.registerModule("products_module", products_module);
